+++ Changelog.md
@@ -19,7 +19,9 @@ Thanks to Patman for providing the download links of the tools !!!
 
 - new: x265 switch --selective-sao added
 - new: L-Smash Works parameters prefer_hw 3 HW auto added to menu
-- new: allow to set the bitrate and size to 0 in the main dialog, nvenc has usecases for it
+- new: the way --vbr-quality works in nvenc was changed,
+       if it's value is -1 the bitrate in the main dialog is used,
+	   if it's higher then -1 the bitrate is set to 0
 - new: add support of video track title/name using MP4Box
 - new: simple QP mode added to nvenc to use one instead of three QP values
 - new: the menu renderer has now Win 10 style arrows for sub menu indication